fix: move title bar on top of tabs on Windows
